# services/booking-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Manage schedules and appointments.",
        "title": "Booking Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8083",
    "basePath": "/",
    "paths": {
        "/bookings": {
            "post": {
                "description": "Create a new appointment if the slot is available",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Book an appointment",
                "parameters": [
                    {
                        "description": "Booking Info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/delivery_http.createBookingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Appointment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/erru.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/erru.AppError"
                        }
                    }
                }
            }
        },
        "/schedules/{employee_id}": {
            "get": {
                "description": "Get weekly working schedule for an employee",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedules"
                ],
                "summary": "Get employee schedule",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "employee_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Schedule"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Update working hours for an employee",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedules"
                ],
                "summary": "Set employee schedule",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "employee_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Schedules Array",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Schedule"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shifts": {
            "get": {
                "description": "Get shifts for an employee or an entire branch for a specific month",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shifts"
                ],
                "summary": "Get work shifts",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "employee_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Branch ID",
                        "name": "branch_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Month (ISO8601, e.g., 2026-01-01T00:00:00Z)",
                        "name": "month",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.WorkShift"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create or update multiple work shifts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shifts"
                ],
                "summary": "Bulk save work shifts",
                "parameters": [
                    {
                        "description": "Shifts Array",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.WorkShift"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/slots": {
            "get": {
                "description": "Calculate available time slots for an employee and service on a specific date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Get available slots",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "employee_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Service ID",
                        "name": "service_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date (ISO8601)",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Slot"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/erru.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/erru.AppError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "delivery_http.createBookingRequest": {
            "type": "object",
            "required": [
                "client_id",
                "employee_id",
                "end_time",
                "service_id",
                "start_time"
            ],
            "properties": {
                "client_id": {
                    "type": "integer"
                },
                "comment": {
                    "type": "string"
                },
                "employee_id": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "service_id": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "domain.Appointment": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "integer"
                },
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "employee_id": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "service_id": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.AppointmentStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.AppointmentStatus": {
            "type": "string",
            "enum": [
                "pending",
                "confirmed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusConfirmed",
                "StatusCancelled"
            ]
        },
        "domain.Schedule": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "day_of_week": {
                    "description": "0 = Sunday, 1 = Monday, ...",
                    "type": "integer"
                },
                "employee_id": {
                    "type": "integer"
                },
                "end_time": {
                    "description": "e.g. \"18:00\"",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_day_off": {
                    "type": "boolean"
                },
                "start_time": {
                    "description": "e.g. \"09:00\"",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.Slot": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "is_free": {
                    "type": "boolean"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "domain.WorkShift": {
            "type": "object",
            "properties": {
                "branch_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "employee_id": {
                    "type": "integer"
                },
                "end_time": {
                    "description": "e.g. \"18:00\"",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_day_off": {
                    "type": "boolean"
                },
                "start_time": {
                    "description": "e.g. \"09:00\"",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "erru.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        }
    }
}